refactor: 상품 선택 로직을 useProductTabState 훅으로 분리
